refactor: 할 일 상세 컴포넌트 분리
